Finished class 6 challenge
Modified the Player class to obtain the correct values to manage a kind of timer to provide a limit to the amount of shoots that can be performed.
Modified the BulletPrototypeController to manage a kind of timer to Destroy the bullets that had been created some time ago.
Modified the BulletPrototypeController to be able to duplicate the size of the bullet when space is pressed and added a bool that allows to be able to resize more than once